combine analyzer into task
